\name{RandomLeter3.pkg-package}
\alias{RandomLeter3.pkg-package}
\alias{RandomLeter3.pkg}
\docType{package}
\title{
\packageTitle{RandomLeter3.pkg}
}
\description{
\packageDescription{RandomLeter3.pkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RandomLeter3.pkg}
\packageIndices{RandomLeter3.pkg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RandomLeter3.pkg}

Maintainer: \packageMaintainer{RandomLeter3.pkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
